+++ submission/helpers.py
@@ -80,7 +80,7 @@ def send_email(subject, reply_to, recipients, text_body, html_body=None):
     message.send()
 
 
-def send_gov_notify(
+def send_gov_notify_email(
     template_id, email_address, personalisation, email_reply_to_id=None
 ):
     client = NotificationsAPIClient(
@@ -91,6 +91,16 @@ def send_gov_notify(
         template_id=template_id,
         personalisation=personalisation,
         email_reply_to_id=email_reply_to_id,
+    )
+
+
+def send_gov_notify_letter(template_id, personalisation):
+    client = NotificationsAPIClient(
+        settings.GOV_NOTIFY_API_KEY,
+    )
+    client.send_letter_notification(
+        template_id=template_id,
+        personalisation=personalisation,
     )
 
 
@@ -108,7 +118,9 @@ def get_sender_email_address(submission_meta):
         return submission_meta['email_address']
     elif action_name == constants.ACTION_NAME_EMAIL:
         return submission_meta['reply_to'][0]
-    elif action_name == constants.ACTION_NAME_GOV_NOTIFY:
+    elif action_name == constants.ACTION_NAME_GOV_NOTIFY_EMAIL:
         return submission_meta['email_address']
     elif action_name == constants.ACTION_NAME_PARDOT:
+        return None
+    elif action_name == constants.ACTION_NAME_GOV_NOTIFY_LETTER:
         return None